backend/pyproject.toml
[tool.poetry]
name = "rr-chatbot-backend"
version = "0.1.0"
description = "Backend for the RR-Chatbot project"
authors = ["vidhya101 <vidhyashankargoel1996@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"  # Pinned to >=3.11,<3.12 for tensorflow-cpu compatibility
typing-extensions = "^4.8"
numpy = "^1.23"
pandas = "^2.0"
scipy = "^1.11"
scikit-learn = "^1.3"
pillow = "^10.1"
pyarrow = "^11.0"
flask = "^2.2"
flask-cors = "^4.0"
flask-sqlalchemy = "^3.0"
tensorflow-cpu = "2.12.0"
tensorflow-io-gcs-filesystem = "0.31.0"  # Added explicitly
keras = "^2.12"
torch = "^2.0"
sqlalchemy = "^1.4"
werkzeug = "^2.2"
alembic = "^1.9"
psycopg2-binary = "^2.9"
redis = "^5.0"
pyjwt = "^2.8"
bcrypt = "^4.1"
python-dotenv = "^1.0"
flask-socketio = "^5.3"
python-socketio = "^5.9"
eventlet = "^0.33"
gevent = "^22.10"
gevent-websocket = "^0.10"
dash = "^2.12"
flask-jwt-extended = "^4.3"
flask-migrate = "^4.0"
requests = "^2.31"
aiohttp = "^3.9"
fastapi = "^0.100"
uvicorn = "^0.23"
pydantic = "^2.5"
torchvision = "^0.15"
pytorch-lightning = "^1.9"
optuna = "^3.5"
shap = "^0.44"
lime = "^0.2"
scikit-image = "^0.22"
imageio = "^2.37"
tifffile = "^2025.3"
lazy-loader = "^0.4"
mlflow = "^2.11"
mistralai = "0.0.12"
openpyxl = "^3.1"
python-docx = "^0.8"
pypdf2 = "^3.0"
python-pptx = "^0.6"
xlrd = "^2.0"
matplotlib = "^3.7"
seaborn = "^0.12"
plotly = "^5.13"
pytest = "^7.3"
black = "^23.3"
flake8 = "^6.0"
pytest-cov = "^4.1"
pytest-mock = "^3.12"
python-dateutil = "^2.9"
pytz = "^2023.3"
psutil = "^5.9"
tqdm = "^4.66"
joblib = "^1.3"
pymongo = "^4.5"
python-magic = "^0.4"
cryptography = "^41.0"
urllib3 = "^1.26"

[tool.poetry.dev-dependencies]
# Add dev dependencies if needed later

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"